Fix code using moduleId as uint32_t type.
Some piece of code are using moduleId as uint32_t type, whereas
moduleId is uint16_t type.

This patch fixes it.

+++ components/Core/include/Core/BaseModelConverter.hpp
@@ -91,10 +91,10 @@ protected:
     static const std::vector<std::string> staticServiceTypes;
 
     /** Find a module entry using its id */
-    const cavs::dsp_fw::ModuleEntry &findModuleEntry(uint32_t moduleId);
+    const cavs::dsp_fw::ModuleEntry &findModuleEntry(uint16_t moduleId);
 
     /** Find a module entry name using its id */
-    std::string findModuleEntryName(uint32_t moduleId);
+    std::string findModuleEntryName(uint16_t moduleId);
 
     /** Find a gateway type name using a connector id */
     static std::string findGatewayTypeName(const cavs::dsp_fw::ConnectorNodeId &connectorId);
